refactor(app): give the tool rail's sections their own siblings
`toolrail.rs` carried 2,114 production lines: the rail's state, the
drawing of every section, the favourites band and the family flyout in
one file. Three sections move out unchanged, each as a file plus one
`mod` line -- `toolrail/state.rs` (the armed tool, the dock, the starred
tools, the band offset, the hit rectangles a script reads and the
keyboard shortcuts), `toolrail/band.rs` (the pinned favourites and the
scrolling band with its chevrons) and `toolrail/flyout.rs` (the family
slot, its flyout and the `FlyoutClick` a row answers with).

Pure moves: every body is unchanged. Five helpers the parent or a
sibling still calls are widened from private to `pub(super)`, which is
what a child module needs and no more, and rustfmt rewrapped one
signature that no longer fits. What stays in `toolrail.rs` is the rail's
own drawing -- the frame, the grip, the separator, the repeat and magnet
toggles, the overflow menu, the objects button and the global toggles.
