+++ crates/app/src/toolrail.rs
@@ -16,6 +16,10 @@ use crate::drawings::{
 use crate::theme;
 use crate::widgets::{IconButton, MarkerEdge, TOOLRAIL_ICON, paint_vector_icon};
 
+mod band;
+mod flyout;
+mod state;
+
 /// Rail cross axis, all four docks: `44 = 6 + 32 + 6`.
 const TOOLBOX_THICKNESS_PX: f32 = 44.0;
 /// Outer margin, both axes.
@@ -659,351 +663,6 @@ impl Default for ToolRail {
 }
 
 impl ToolRail {
-    #[must_use]
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    #[must_use]
-    pub fn tool(&self) -> Tool {
-        self.tool
-    }
-
-    #[must_use]
-    pub fn visible(&self) -> bool {
-        self.visible
-    }
-
-    #[must_use]
-    pub fn dock(&self) -> ToolboxDock {
-        self.dock
-    }
-
-    /// Dock the rail against `dock` — the menu path beside dragging.
-    pub fn set_dock(&mut self, dock: ToolboxDock) {
-        self.dock = dock;
-    }
-
-    /// Show or hide the rail outright, for a saved workspace restoring the
-    /// state it recorded rather than toggling from whatever this launch
-    /// happens to be in.
-    pub fn set_visible(&mut self, visible: bool) {
-        if self.visible != visible {
-            self.toggle_visible();
-        }
-    }
-
-    /// The controls the rail painted last frame, in rail order.
-    ///
-    /// Folded through the very `tool_slots()` the draw folds through, and cut
-    /// by the stage and the band window the draw recorded, so this can never
-    /// name a button that is behind a flyout, off the scrolled run, or inside
-    /// the More menu. A rail nobody can see, and one that has not been drawn
-    /// yet, both report nothing: the guarantee is made here rather than left
-    /// to whichever caller remembers to ask [`Self::visible`] first.
-    #[must_use]
-    pub(crate) fn painted_controls(&self) -> Vec<RailControl> {
-        if !self.visible {
-            return Vec::new();
-        }
-        let Some(stage) = self.last_stage else {
-            return Vec::new();
-        };
-        let armed_drawing = self.tool.drawing_tool();
-        let mut controls = vec![RailControl {
-            kind: RailControlKind::Tool,
-            id: Tool::Pointer.id(),
-            label: Tool::Pointer.name(),
-            armed: self.tool == Tool::Pointer,
-        }];
-        // Minimal drops the crosshair; every wider stage keeps it.
-        if stage != RailStage::Minimal {
-            controls.push(RailControl {
-                kind: RailControlKind::Tool,
-                id: Tool::Crosshair.id(),
-                label: Tool::Crosshair.name(),
-                armed: self.tool == Tool::Crosshair,
-            });
-        }
-        match stage {
-            // The whole run has a button, and so does every star: nothing is
-            // clipped and nothing is folded away at this width.
-            RailStage::Full => {
-                for tool in &self.favorites {
-                    controls.push(self.favorite_control(*tool));
-                }
-                for slot in tool_slots() {
-                    controls.push(Self::slot_control(slot, armed_drawing));
-                }
-            }
-            // The band clips. Only the anchored stars stay outside it; the
-            // rest of the stars and the whole tool run scroll behind two
-            // chevrons, and only the window the draw recorded is on screen.
-            RailStage::Scroll => {
-                let Some(band) = self.last_band else {
-                    return controls;
-                };
-                let anchored = band.anchored.min(self.favorites.len());
-                for tool in &self.favorites[..anchored] {
-                    controls.push(self.favorite_control(*tool));
-                }
-                let slots = tool_slots();
-                let spilled = self.favorites.len() - anchored;
-                let content = spilled + slots.len();
-                let first = band.first.min(content);
-                let last = first.saturating_add(band.visible).min(content);
-                for index in first..last {
-                    controls.push(match index.checked_sub(spilled) {
-                        Some(slot) => Self::slot_control(&slots[slot], armed_drawing),
-                        None => self.favorite_control(self.favorites[anchored + index]),
-                    });
-                }
-            }
-            // Only the armed tool keeps a button of its own; the rest are
-            // behind More, which is not a control the scene can name yet.
-            // The stars go with them — these stages draw no pinned section.
-            RailStage::Compact | RailStage::Minimal => {
-                if let Some(tool) = armed_drawing {
-                    controls.push(RailControl {
-                        kind: RailControlKind::Tool,
-                        id: tool.id(),
-                        label: tool.name(),
-                        armed: true,
-                    });
-                }
-            }
-        }
-        controls
-    }
-
-    /// One entry of the folded tool run, as the draw paints it.
-    fn slot_control(slot: &RailSlot, armed_drawing: Option<DrawingTool>) -> RailControl {
-        match slot {
-            RailSlot::Single(tool) => RailControl {
-                kind: RailControlKind::Tool,
-                id: tool.id(),
-                label: tool.name(),
-                armed: armed_drawing == Some(*tool),
-            },
-            RailSlot::Family { family, members } => RailControl {
-                kind: RailControlKind::Family,
-                id: family.id,
-                label: family.title,
-                armed: armed_drawing.is_some_and(|tool| members.contains(&tool)),
-            },
-        }
-    }
-
-    /// One pinned button. A star is a second button for a tool that also has
-    /// a slot in the run, so it is named in its own namespace rather than
-    /// answering to the slot's name.
-    fn favorite_control(&self, tool: DrawingTool) -> RailControl {
-        RailControl {
-            kind: RailControlKind::Favorite,
-            id: tool.id(),
-            label: tool.name(),
-            armed: self.tool == Tool::Drawing(tool),
-        }
-    }
-
-    /// Whether a grip drag is live this frame — the app's escape stack must
-    /// yield Esc to the drag while it is.
-    #[must_use]
-    pub fn drag_active(&self) -> bool {
-        self.dragging
-    }
-
-    pub fn toggle_visible(&mut self) {
-        self.visible = !self.visible;
-        if !self.visible {
-            self.tool = Tool::Pointer;
-            // Nothing is painted while the rail is away, so what the last
-            // draw settled on stops describing the screen. Forgotten rather
-            // than kept: control captures are served before the rail draws,
-            // so a stale stage would answer the first capture after a hide
-            // with the buttons of a rail nobody can see.
-            self.last_stage = None;
-            self.last_band = None;
-        }
-    }
-
-    pub fn arm(&mut self, tool: Tool) {
-        if let Tool::Drawing(drawing_tool) = tool
-            && let Some(family) = drawing_tool.family()
-        {
-            self.last_family_member.insert(family.id, drawing_tool);
-        }
-        if self.tool != tool {
-            self.reveal_armed = true;
-        }
-        self.tool = tool;
-    }
-
-    /// The starred tools, in the order they were starred.
-    #[must_use]
-    pub fn favorites(&self) -> &[DrawingTool] {
-        &self.favorites
-    }
-
-    #[must_use]
-    pub fn is_favorite(&self, tool: DrawingTool) -> bool {
-        self.favorites.contains(&tool)
-    }
-
-    /// Star or unstar a tool. Starring appends — the pinned section grows at
-    /// its far end, and the tools already pinned never move.
-    pub fn toggle_favorite(&mut self, tool: DrawingTool) {
-        if let Some(index) = self.favorites.iter().position(|entry| *entry == tool) {
-            self.favorites.remove(index);
-        } else {
-            self.favorites.push(tool);
-        }
-        self.favorites_changed = true;
-    }
-
-    /// Whether the trader starred or unstarred something since this was last
-    /// asked, clearing the flag.
-    ///
-    /// The rail curates the list; it does not know where lists are kept. The
-    /// app reads this on the frame the star was clicked and writes the choice
-    /// down — the same hand-off the replay browser's folder pick uses, and for
-    /// the same reason: "it forgot my tools again" must not be one crash away.
-    pub fn take_favorites_change(&mut self) -> bool {
-        std::mem::take(&mut self.favorites_changed)
-    }
-
-    /// Restore the starred list from saved tool ids — the workspace file
-    /// path. An id no registered tool carries is dropped, a duplicate keeps
-    /// its first position, and saved order is kept: it is the order the
-    /// trader starred in.
-    ///
-    /// Returns the ids it did not recognise, so the caller can say so. The
-    /// prune used to be private and harmless — it only reached the disk on an
-    /// explicit save. Now the trader's next star click writes this list back,
-    /// which makes the loss permanent, and losing a saved id without a word is
-    /// exactly the silent patching `CLAUDE.md` rules out. The sibling restore
-    /// logs `UI_STATE_TAB_DROPPED` for a tab it cannot open, for the same
-    /// reason.
-    pub fn set_favorites(&mut self, ids: &[String]) -> Vec<String> {
-        self.favorites.clear();
-        let mut unknown = Vec::new();
-        for id in ids {
-            match DrawingTool::by_id(id) {
-                Some(tool) if !self.favorites.contains(&tool) => self.favorites.push(tool),
-                // A duplicate is not a loss: the first position stands.
-                Some(_) => {}
-                None => unknown.push(id.clone()),
-            }
-        }
-        unknown
-    }
-
-    /// Park the scrolling tool band at `offset` px along the rail — the
-    /// validation hook for a state that otherwise takes a chevron click.
-    /// The band clamps it on the next frame, so `f32::INFINITY` means "the
-    /// far end" and a rail that does not scroll ignores it entirely.
-    pub fn set_band_offset(&mut self, offset: f32) {
-        self.band_target = Some(offset.max(0.0));
-    }
-
-    /// Whether the repeat pin keeps the tool armed after an object completes.
-    #[must_use]
-    pub fn repeat(&self) -> bool {
-        self.repeat
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_repeat(&mut self, repeat: bool) {
-        self.repeat = repeat;
-    }
-
-    /// Whether placed anchors snap to the bar's open / high / low / close.
-    #[must_use]
-    pub fn magnet(&self) -> bool {
-        self.magnet
-    }
-
-    /// Arm the magnet without a click — the `QUANTICK_DRAWING_MAGNET` hook
-    /// and the tests both come through here, so neither can drift from what
-    /// the button does.
-    pub(crate) fn set_magnet(&mut self, magnet: bool) {
-        self.magnet = magnet;
-    }
-
-    /// Ask for a family flyout without a click — the
-    /// `QUANTICK_TOOLBOX_FLYOUT` hook. The slot honours it on its next draw,
-    /// when the anchor rect exists; an unknown family id is simply never
-    /// matched and stays pending, which draws nothing.
-    pub(crate) fn request_flyout(&mut self, family_id: String) {
-        self.hook_flyout = Some(family_id);
-    }
-
-    #[cfg(test)]
-    pub(crate) fn button_rect(&self, tool: Tool) -> Option<egui::Rect> {
-        self.button_rects
-            .iter()
-            .flatten()
-            .find_map(|(candidate, rect)| (*candidate == tool).then_some(*rect))
-    }
-
-    #[cfg(test)]
-    pub(crate) fn objects_button_rect(&self) -> Option<egui::Rect> {
-        self.objects_rect
-    }
-
-    #[cfg(test)]
-    pub(crate) fn flyout_row_rect(&self, tool: DrawingTool) -> Option<egui::Rect> {
-        self.flyout_rects
-            .iter()
-            .find_map(|(candidate, rect)| (*candidate == tool).then_some(*rect))
-    }
-
-    /// Tool-arming keys. Escape lives in the app's escape stack (rail drag →
-    /// input → draft → selection → Pointer), not here. Each drawing tool
-    /// declares its own shortcut through the registry.
-    pub fn handle_keys(&mut self, ctx: &egui::Context) {
-        // A hidden rail arms nothing (audit M9): with no rail on screen an
-        // armed tool has no indication anywhere, and the next chart click
-        // would draw instead of pan — the keyboard twin of the invariant
-        // `hiding_the_toolbox_cannot_leave_an_invisible_drawing_tool_armed`.
-        if !self.visible {
-            return;
-        }
-        if ctx.memory(|memory| memory.focused().is_some()) {
-            return;
-        }
-        let armed = ctx.input(|input| {
-            if input.modifiers.command || input.modifiers.alt {
-                return None;
-            }
-            if input.key_pressed(egui::Key::Num1) {
-                return Some(Tool::Pointer);
-            }
-            if input.key_pressed(egui::Key::Num2) {
-                return Some(Tool::Crosshair);
-            }
-            DRAWING_TOOLS.into_iter().find_map(|tool| {
-                tool.shortcut()
-                    .filter(|shortcut| {
-                        // A tool's shortcut is a *bare* key (with Shift where
-                        // the tool asks for it). Ctrl+M, Cmd+M and Alt+M
-                        // belong to whoever claims them — the mark hotkey does
-                        // — and must not also arm a tool out from under the
-                        // trader's hand.
-                        input.key_pressed(shortcut.key)
-                            && input.modifiers.shift == shortcut.shift
-                            && !input.modifiers.ctrl
-                            && !input.modifiers.command
-                            && !input.modifiers.alt
-                    })
-                    .map(|_| Tool::Drawing(tool))
-            })
-        });
-        if let Some(tool) = armed {
-            self.arm(tool);
-        }
-    }
-
     /// Draw the rail docked against its edge. Drag the grip and release: the
     /// nearest window edge becomes the new dock, previewed live by a band.
     /// The rail also hosts the object-manager entry and the global protection
@@ -1473,578 +1132,6 @@ impl ToolRail {
         paint_badge(ui, response.rect, &text, theme::ACCENT);
         true
     }
-
-    /// The pinned section at the tool end of the rail: a separator, then one
-    /// button per starred tool in star order. One click arms; unstarring
-    /// lives only on the star in the tool's own flyout row, so a pinned
-    /// button can never be destroyed by the click meant to use it.
-    fn draw_favorites_section(
-        &mut self,
-        ui: &mut egui::Ui,
-        vertical: bool,
-        drawings: &Drawings,
-        range: std::ops::Range<usize>,
-    ) {
-        if range.is_empty() {
-            return;
-        }
-        self.draw_separator(ui, vertical);
-        self.draw_favorite_buttons(ui, drawings, range);
-    }
-
-    /// The pinned buttons themselves, without the section separator — the
-    /// band reuses this for favorites that spilled past the anchored head.
-    fn draw_favorite_buttons(
-        &mut self,
-        ui: &mut egui::Ui,
-        drawings: &Drawings,
-        range: std::ops::Range<usize>,
-    ) {
-        // Indexed so the loop never clones the list on the per-frame path;
-        // `DrawingTool` is `Copy` and arming cannot reorder favorites.
-        for index in range {
-            let tool = self.favorites[index];
-            let armed = self.tool == Tool::Drawing(tool);
-            let response = IconButton::new(tool.icon(), TOOLRAIL_ICON)
-                .vector_icon(tool.icon_strokes(), tool.icon_dots(), tool.icon_letter())
-                .active(armed)
-                .active_marker(self.dock.marker_edge())
-                .hover_text(tool.hover_text())
-                .show(ui);
-            self.paint_draft_badge(ui, &response, Tool::Drawing(tool), drawings);
-            if ui.is_rect_visible(response.rect) {
-                // The corner star names the section: this button is a pin,
-                // not a second registry slot.
-                ui.painter().text(
-                    egui::pos2(
-                        response.rect.right() - FAVORITE_BADGE_INSET_PX,
-                        response.rect.top() + FAVORITE_BADGE_INSET_PX,
-                    ),
-                    egui::Align2::RIGHT_TOP,
-                    icons::STAR,
-                    egui::FontId::proportional(FAVORITE_BADGE_PX),
-                    theme::ACCENT,
-                );
-            }
-            #[cfg(test)]
-            self.favorite_rects.push((tool, response.rect));
-            if response.clicked() {
-                self.arm(Tool::Drawing(tool));
-            }
-        }
-    }
-
-    /// The scrolling tool band: a chevron at each end and, between them, the
-    /// favorites that spilled past the anchored head followed by every tool
-    /// slot. Nothing leaves the inventory — the band is a window onto it, so
-    /// a fourth star can no longer swallow the toolbar.
-    fn draw_band(
-        &mut self,
-        ui: &mut egui::Ui,
-        vertical: bool,
-        drawings: &Drawings,
-        slots: &'static [RailSlot],
-        anchored: usize,
-        available: f32,
-    ) {
-        let viewport = band_viewport(available, anchored);
-        let spilled = self.favorites.len() - anchored;
-        let max_offset = band_max_offset(viewport, spilled + slots.len());
-        // Arming a tool the band has scrolled past pulls it back into view,
-        // so the rail keeps the spec's promise that the armed tool always
-        // has a real slot. Only on the frame it was armed: past that, the
-        // band stays wherever the trader put it.
-        if self.reveal_armed {
-            self.reveal_armed = false;
-            if let Some(index) = self.armed_band_index(slots, spilled) {
-                let span = TOOLRAIL_ICON.hit + TOOLBOX_ITEM_GAP_PX;
-                let visible = band_visible_items(viewport);
-                let first = (self.band_offset / span).round() as usize;
-                if index < first {
-                    self.band_target = Some(index as f32 * span);
-                } else if index >= first + visible {
-                    self.band_target = Some((index + 1 - visible) as f32 * span);
-                }
-            }
-        }
-        // A pending chevron click is resolved before anything is drawn, so
-        // both chevrons and the band read one offset. Taking it from last
-        // frame instead would leave the way-back arrow a frame stale — dead
-        // on the very click that opened it.
-        let target = self.band_target.take().map(|at| at.clamp(0.0, max_offset));
-        let offset = target.unwrap_or(self.band_offset).clamp(0.0, max_offset);
-        self.band_offset = offset;
-        let step = band_scroll_step(viewport);
-
-        if self.draw_chevron(ui, vertical, true, offset > 0.0) {
-            self.band_target = Some(offset - step);
-        }
-
-        let mut area = egui::ScrollArea::new([!vertical, vertical])
-            .id_salt("toolrail_band")
-            .scroll_bar_visibility(egui::scroll_area::ScrollBarVisibility::AlwaysHidden)
-            .auto_shrink([false, false]);
-        area = if vertical {
-            area.max_height(viewport)
-        } else {
-            area.max_width(viewport)
-        };
-        if let Some(target) = target {
-            area = if vertical {
-                area.vertical_scroll_offset(target)
-            } else {
-                area.horizontal_scroll_offset(target)
-            };
-        }
-        let inner = if vertical {
-            egui::Layout::top_down(egui::Align::Center)
-        } else {
-            egui::Layout::left_to_right(egui::Align::Center)
-        };
-        // The band is handed its extent rather than left to ask for one:
-        // told only a maximum, a scroll area still claims whatever the
-        // parent has left, and the far cluster needs that room. Its bar is
-        // also denied a lane across the rail's short axis, because a hidden
-        // bar still books one and 44 px has none to spare.
-        let band_size = if vertical {
-            egui::vec2(ui.available_width(), viewport)
-        } else {
-            egui::vec2(viewport, ui.available_height())
-        };
-        let output = ui
-            .allocate_ui_with_layout(band_size, inner, |ui| {
-                let scroll = &mut ui.spacing_mut().scroll;
-                scroll.floating = true;
-                scroll.bar_width = 0.0;
-                scroll.floating_allocated_width = 0.0;
-                area.show(ui, |ui| {
-                    ui.with_layout(inner, |ui| {
-                        ui.spacing_mut().item_spacing =
-                            egui::vec2(TOOLBOX_ITEM_GAP_PX, TOOLBOX_ITEM_GAP_PX);
-                        let favorites = self.favorites.len();
-                        self.draw_favorite_buttons(ui, drawings, anchored..favorites);
-                        for slot in slots {
-                            match slot {
-                                RailSlot::Single(tool) => {
-                                    self.draw_button(ui, Tool::Drawing(*tool), drawings);
-                                }
-                                RailSlot::Family { family, members } => {
-                                    self.draw_family_slot(ui, *family, members, drawings);
-                                }
-                            }
-                        }
-                    });
-                })
-            })
-            .inner;
-        #[cfg(test)]
-        {
-            self.band_rect = Some(output.inner_rect);
-        }
-        // The wheel moves the band without asking us, so the offset is read
-        // back rather than assumed.
-        let scrolled = if vertical {
-            output.state.offset.y
-        } else {
-            output.state.offset.x
-        };
-        self.band_offset = scrolled.clamp(0.0, max_offset);
-        // The window this draw actually showed, written down for the same
-        // reason the stage is: a reader that is not looking at the screen
-        // cannot re-derive it, because it depends on the extent the layout
-        // handed the rail and on wherever the trader last scrolled to.
-        let span = TOOLRAIL_ICON.hit + TOOLBOX_ITEM_GAP_PX;
-        self.last_band = Some(BandWindow {
-            anchored,
-            first: (self.band_offset / span).round() as usize,
-            visible: band_visible_items(viewport),
-        });
-
-        if self.draw_chevron(ui, vertical, false, self.band_offset < max_offset) {
-            self.band_target = Some(self.band_offset + step);
-        }
-        // The trailing chevron sits past the band, so its click can only be
-        // honoured next frame. Ask for that frame: without it a click on a
-        // still chart reads as a dead button.
-        if self.band_target.is_some() {
-            ui.ctx().request_repaint();
-        }
-    }
-
-    /// Where the armed tool sits among the band's items: the spilled
-    /// favorites first, then the tool slots in registry order. `None` when
-    /// nothing is armed, or when the armed tool is anchored outside the band
-    /// and therefore already on screen.
-    fn armed_band_index(&self, slots: &[RailSlot], spilled: usize) -> Option<usize> {
-        let armed = self.tool.drawing_tool()?;
-        let anchored = self.favorites.len() - spilled;
-        if let Some(offset) = self.favorites[anchored..]
-            .iter()
-            .position(|pinned| *pinned == armed)
-        {
-            return Some(offset);
-        }
-        let slot = slots.iter().position(|slot| match slot {
-            RailSlot::Single(tool) => *tool == armed,
-            RailSlot::Family { members, .. } => members.contains(&armed),
-        })?;
-        Some(spilled + slot)
-    }
-
-    /// One end of the band's navigation pair. Both ends keep their slot for
-    /// as long as the band scrolls: a chevron that vanished at the end of
-    /// travel would shift every tool under the pointer by its own length.
-    /// A dead end dims and stops sensing clicks instead.
-    fn draw_chevron(
-        &mut self,
-        ui: &mut egui::Ui,
-        vertical: bool,
-        leading: bool,
-        live: bool,
-    ) -> bool {
-        let size = if vertical {
-            egui::vec2(TOOLRAIL_ICON.hit, BAND_ARROW_LENGTH_PX)
-        } else {
-            egui::vec2(BAND_ARROW_LENGTH_PX, TOOLRAIL_ICON.hit)
-        };
-        let sense = if live {
-            egui::Sense::click()
-        } else {
-            egui::Sense::hover()
-        };
-        let (rect, response) = ui.allocate_exact_size(size, sense);
-        #[cfg(test)]
-        if leading {
-            self.band_leading_arrow = Some((rect, live));
-        } else {
-            self.band_trailing_arrow = Some((rect, live));
-        }
-        if ui.is_rect_visible(rect) {
-            let glyph = match (vertical, leading) {
-                (true, true) => icons::CARET_UP,
-                (true, false) => icons::CARET_DOWN,
-                (false, true) => icons::CARET_LEFT,
-                (false, false) => icons::CARET_RIGHT,
-            };
-            let color = if !live {
-                theme::TEXT_FAINT
-            } else if response.hovered() {
-                theme::TEXT_PRIMARY
-            } else {
-                theme::TEXT_MUTED
-            };
-            ui.painter().text(
-                rect.center(),
-                egui::Align2::CENTER_CENTER,
-                glyph,
-                egui::FontId::proportional(BAND_ARROW_GLYPH_PX),
-                color,
-            );
-        }
-        // A dimmed control with no reason reads as a bug, so the dead end
-        // says which end it is rather than saying nothing.
-        let hint = match (vertical, leading, live) {
-            (true, true, true) => "Scroll tools up",
-            (true, false, true) => "Scroll tools down",
-            (false, true, true) => "Scroll tools left",
-            (false, false, true) => "Scroll tools right",
-            (_, true, false) => "Start of the tool list",
-            (_, false, false) => "End of the tool list",
-        };
-        live && response.on_hover_text(hint).clicked()
-    }
-
-    /// A family's shown member: the last-armed one, or `None` before any
-    /// member has been used.
-    fn family_member(&self, family: ToolFamily, members: &[DrawingTool]) -> Option<DrawingTool> {
-        self.last_family_member
-            .get(family.id)
-            .copied()
-            .filter(|member| members.contains(member))
-    }
-
-    /// The family split button: left-click arms the shown member; the caret
-    /// zone or a right-click opens the member flyout.
-    fn draw_family_slot(
-        &mut self,
-        ui: &mut egui::Ui,
-        family: ToolFamily,
-        members: &[DrawingTool],
-        drawings: &Drawings,
-    ) {
-        let shown = self.family_member(family, members);
-        let armed = self
-            .tool
-            .drawing_tool()
-            .is_some_and(|tool| members.contains(&tool));
-        let icon = shown.map_or(family.icon, DrawingTool::icon);
-        let strokes = shown.map_or(family.icon_strokes, DrawingTool::icon_strokes);
-        let dots = shown.map_or(family.icon_dots, DrawingTool::icon_dots);
-        let letter = shown.map_or(family.icon_letter, DrawingTool::icon_letter);
-        let hover = shown.map_or(family.title, DrawingTool::hover_text);
-        let response = IconButton::new(icon, TOOLRAIL_ICON)
-            .vector_icon(strokes, dots, letter)
-            .active(armed)
-            .active_marker(self.dock.marker_edge())
-            .hover_text(hover)
-            .show(ui);
-        let badge_shown = shown
-            .map(Tool::Drawing)
-            .is_some_and(|tool| self.paint_draft_badge(ui, &response, tool, drawings));
-
-        // The caret marks the flyout; it yields the corner to a draft badge,
-        // because a tool mid-draft is unambiguously armed already.
-        let caret_zone = egui::Rect::from_min_max(
-            egui::pos2(
-                response.rect.right() - TOOLBOX_CARET_ZONE_PX,
-                response.rect.bottom() - TOOLBOX_CARET_ZONE_PX,
-            ),
-            response.rect.max,
-        );
-        if !badge_shown && ui.is_rect_visible(response.rect) {
-            let caret_color = if armed {
-                theme::ACCENT
-            } else if response.hovered() {
-                theme::TEXT_PRIMARY
-            } else {
-                theme::TEXT_FAINT
-            };
-            let corner = egui::pos2(
-                response.rect.right() - CARET_INSET_PX,
-                response.rect.bottom() - CARET_INSET_PX,
-            );
-            ui.painter().add(egui::Shape::convex_polygon(
-                vec![
-                    egui::pos2(corner.x - CARET_SIDE_PX, corner.y),
-                    corner,
-                    egui::pos2(corner.x, corner.y - CARET_SIDE_PX),
-                ],
-                caret_color,
-                egui::Stroke::NONE,
-            ));
-        }
-
-        #[cfg(test)]
-        if let Some(slot) = self.button_rects.iter_mut().find(|slot| slot.is_none()) {
-            let recorded = shown.unwrap_or(members[0]);
-            *slot = Some((Tool::Drawing(recorded), response.rect));
-        }
-
-        if self.hook_flyout.as_deref() == Some(family.id) {
-            self.hook_flyout = None;
-            self.flyout = Some((family.id, response.rect));
-        }
-
-        let caret_clicked = response.clicked()
-            && response
-                .interact_pointer_pos()
-                .is_some_and(|position| caret_zone.contains(position));
-        if response.secondary_clicked() || caret_clicked {
-            self.flyout = Some((family.id, response.rect));
-        } else if response.clicked() {
-            self.arm(Tool::Drawing(shown.unwrap_or(members[0])));
-        }
-    }
-
-    /// The open family flyout, on the rail's chart-facing side, first row
-    /// aligned with the slot's leading edge.
-    fn draw_family_flyout(&mut self, ctx: &egui::Context) {
-        let Some((family_id, anchor)) = self.flyout else {
-            return;
-        };
-        let Some((family, members)) = tool_slots().iter().find_map(|slot| match slot {
-            RailSlot::Family { family, members } if family.id == family_id => {
-                Some((*family, members.as_slice()))
-            }
-            _ => None,
-        }) else {
-            self.flyout = None;
-            return;
-        };
-
-        if ctx.input(|input| input.key_pressed(egui::Key::Escape)) {
-            self.flyout = None;
-            return;
-        }
-
-        let screen = ctx.screen_rect();
-        let height =
-            TOOLBOX_FLYOUT_ROW_HEIGHT_PX * (members.len() + 1) as f32 + 2.0 * TOOLBOX_ITEM_GAP_PX;
-        let position = match self.dock {
-            ToolboxDock::Left => egui::pos2(anchor.right() + TOOLBOX_MARGIN_PX, anchor.top()),
-            ToolboxDock::Top => egui::pos2(anchor.left(), anchor.bottom() + TOOLBOX_MARGIN_PX),
-            ToolboxDock::Bottom => {
-                egui::pos2(anchor.left(), anchor.top() - TOOLBOX_MARGIN_PX - height)
-            }
-        };
-        let max_position = egui::pos2(
-            (screen.right() - TOOLBOX_FLYOUT_WIDTH_PX).max(screen.left()),
-            (screen.bottom() - height).max(screen.top()),
-        );
-        let position = position.clamp(screen.min, max_position);
-
-        let area = egui::Area::new(egui::Id::new("toolbox_family_flyout"))
-            .order(egui::Order::Foreground)
-            .fixed_pos(position)
-            .show(ctx, |ui| {
-                egui::Frame::popup(ui.style())
-                    .fill(theme::CONTROL)
-                    .stroke(egui::Stroke::new(1.0_f32, theme::BORDER))
-                    .rounding(egui::Rounding::same(FLYOUT_CORNER_RADIUS_PX))
-                    .show(ui, |ui| {
-                        ui.set_width(TOOLBOX_FLYOUT_WIDTH_PX - 2.0 * TOOLBOX_ITEM_GAP_PX);
-                        ui.label(
-                            egui::RichText::new(family.title)
-                                .color(theme::TEXT_MUTED)
-                                .size(FLYOUT_HEADER_TEXT_PX),
-                        );
-                        for member in members {
-                            match self.draw_flyout_row(ui, *member) {
-                                Some(FlyoutClick::Arm) => {
-                                    self.arm(Tool::Drawing(*member));
-                                    self.flyout = None;
-                                }
-                                // Starring neither arms nor closes: the
-                                // trader is curating the rail, not drawing.
-                                Some(FlyoutClick::ToggleFavorite) => {
-                                    self.toggle_favorite(*member);
-                                }
-                                None => {}
-                            }
-                        }
-                    });
-            });
-
-        // A press anywhere outside closes the flyout without arming.
-        if self.flyout.is_some() {
-            let pressed_outside = ctx.input(|input| {
-                input.pointer.any_pressed()
-                    && input.pointer.interact_pos().is_some_and(|position| {
-                        !area.response.rect.contains(position) && !anchor.contains(position)
-                    })
-            });
-            if pressed_outside {
-                self.flyout = None;
-            }
-        }
-    }
-
-    /// One flyout row: glyph with its favorite star, name, right-aligned
-    /// shortcut. Returns what the click asked for, `None` when nothing was
-    /// clicked.
-    fn draw_flyout_row(&mut self, ui: &mut egui::Ui, member: DrawingTool) -> Option<FlyoutClick> {
-        let width = ui.available_width();
-        let (rect, response) = ui.allocate_exact_size(
-            egui::vec2(width, TOOLBOX_FLYOUT_ROW_HEIGHT_PX),
-            egui::Sense::click(),
-        );
-        // The star holds the row's right edge. Its hit zone outranks the
-        // row: a click there curates favorites, everywhere else arms.
-        let star_center = egui::pos2(rect.right() - FLYOUT_STAR_RIGHT_INSET_PX, rect.center().y);
-        let star_zone =
-            egui::Rect::from_center_size(star_center, egui::Vec2::splat(FLYOUT_STAR_HIT_PX));
-        let favorite = self.is_favorite(member);
-        #[cfg(test)]
-        self.flyout_rects.push((member, rect));
-        #[cfg(test)]
-        self.flyout_star_rects.push((member, star_zone));
-        if ui.is_rect_visible(rect) {
-            let armed = self.tool == Tool::Drawing(member);
-            if armed {
-                ui.painter().rect_filled(
-                    rect,
-                    egui::Rounding::same(FLYOUT_ROW_RADIUS_PX),
-                    theme::active_tint(theme::ACCENT),
-                );
-            } else if response.hovered() {
-                ui.painter().rect_filled(
-                    rect,
-                    egui::Rounding::same(FLYOUT_ROW_RADIUS_PX),
-                    theme::BORDER,
-                );
-            }
-            let glyph_color = if armed {
-                theme::ACCENT
-            } else {
-                theme::TEXT_MUTED
-            };
-            let glyph_center = egui::pos2(rect.left() + FLYOUT_GLYPH_CENTER_X_PX, rect.center().y);
-            if member.icon_strokes().is_empty() {
-                ui.painter().text(
-                    glyph_center,
-                    egui::Align2::CENTER_CENTER,
-                    member.icon(),
-                    egui::FontId::proportional(FLYOUT_GLYPH_PX),
-                    glyph_color,
-                );
-            } else {
-                paint_vector_icon(
-                    ui.painter(),
-                    egui::Rect::from_center_size(
-                        glyph_center,
-                        egui::Vec2::splat(FLYOUT_ICON_BOX_PX),
-                    ),
-                    member.icon_strokes(),
-                    member.icon_dots(),
-                    member.icon_letter(),
-                    glyph_color,
-                );
-            }
-            // Accent when starred; otherwise the star only whispers on row
-            // hover, so an unstarred flyout stays as quiet as before.
-            if favorite || response.hovered() {
-                let star_color = if favorite {
-                    theme::ACCENT
-                } else {
-                    theme::TEXT_FAINT
-                };
-                ui.painter().text(
-                    star_center,
-                    egui::Align2::CENTER_CENTER,
-                    icons::STAR,
-                    egui::FontId::proportional(FLYOUT_STAR_PX),
-                    star_color,
-                );
-            }
-            ui.painter().text(
-                egui::pos2(rect.left() + FLYOUT_NAME_X_PX, rect.center().y),
-                egui::Align2::LEFT_CENTER,
-                member.name(),
-                egui::FontId::proportional(FLYOUT_NAME_TEXT_PX),
-                theme::TEXT_PRIMARY,
-            );
-            if let Some(shortcut) = Tool::Drawing(member).shortcut_label() {
-                ui.painter().text(
-                    egui::pos2(
-                        rect.right() - FLYOUT_STAR_SLOT_PX - FLYOUT_SHORTCUT_INSET_PX,
-                        rect.center().y,
-                    ),
-                    egui::Align2::RIGHT_CENTER,
-                    shortcut,
-                    egui::FontId::proportional(FLYOUT_SHORTCUT_TEXT_PX),
-                    theme::TEXT_FAINT,
-                );
-            }
-        }
-        if response.clicked() {
-            let on_star = response
-                .interact_pointer_pos()
-                .is_some_and(|position| star_zone.contains(position));
-            if on_star {
-                return Some(FlyoutClick::ToggleFavorite);
-            }
-            return Some(FlyoutClick::Arm);
-        }
-        None
-    }
-}
-
-/// What a click on a flyout row asked for.
-enum FlyoutClick {
-    /// Arm the row's tool and close the flyout.
-    Arm,
-    /// Star or unstar the row's tool; the flyout stays open.
-    ToggleFavorite,
 }
 
 /// The rail's frame: chrome fill, margins, and no stroke — the chart-facing
